python/faas-driver/requirements.txt
aiohttp==3.8.3 ; python_version >= "3.8" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.8" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.8" and python_version < "4.0"
attrs==22.1.0 ; python_version >= "3.8" and python_version < "4.0"
cachetools==5.2.0 ; python_version >= "3.8" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.8" and python_version < "4.0"
charset-normalizer==2.1.1 ; python_version >= "3.8" and python_version < "4"
click==8.1.3 ; python_version >= "3.8" and python_version < "4.0"
cloudevents==1.8.0 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
dash-core-components==2.0.0 ; python_version >= "3.8" and python_version < "4.0"
dash-html-components==2.0.0 ; python_version >= "3.8" and python_version < "4.0"
dash-table==5.0.0 ; python_version >= "3.8" and python_version < "4.0"
dash==2.7.1 ; python_version >= "3.8" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.8" and python_version < "4.0"
durationpy==0.5 ; python_version >= "3.8" and python_version < "4.0"
flask==2.2.2 ; python_version >= "3.8" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.8" and python_version < "4.0"
google-auth==2.15.0 ; python_version >= "3.8" and python_version < "4.0"
idna==3.4 ; python_version >= "3.8" and python_version < "4"
importlib-metadata==5.1.0 ; python_version >= "3.8" and python_version < "3.10"
inflection==0.5.1 ; python_version >= "3.8" and python_version < "4.0"
iso8601==1.1.0 ; python_version >= "3.8" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.8" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.8" and python_version < "4.0"
kopf==1.36.0 ; python_version >= "3.8" and python_version < "4.0"
kubernetes==25.3.0 ; python_version >= "3.8" and python_version < "4.0"
markupsafe==2.1.1 ; python_version >= "3.8" and python_version < "4.0"
multidict==6.0.3 ; python_version >= "3.8" and python_version < "4.0"
numpy==1.23.5 ; python_version < "4.0" and python_version >= "3.8"
oauthlib==3.2.2 ; python_version >= "3.8" and python_version < "4.0"
packaging==22.0 ; python_version >= "3.8" and python_version < "4.0"
pandas==1.5.2 ; python_version >= "3.8" and python_version < "4.0"
parliament-functions==0.1.0 ; python_version >= "3.8" and python_version < "4.0"
plotly==5.11.0 ; python_version >= "3.8" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.8" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.8" and python_version < "4.0"
python-box==6.1.0 ; python_version >= "3.8" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "4.0"
python-json-logger==2.0.4 ; python_version >= "3.8" and python_version < "4.0"
pytz==2022.6 ; python_version >= "3.8" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.8" and python_version < "4.0"
pyzed==1.3.0 ; python_version >= "3.8" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.8" and python_version < "4.0"
requests==2.28.1 ; python_version >= "3.8" and python_version < "4"
rsa==4.9 ; python_version >= "3.8" and python_version < "4"
setuptools==65.6.3 ; python_version >= "3.8" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
tenacity==8.1.0 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.8" and python_version < "4.0"
urllib3==1.26.13 ; python_version >= "3.8" and python_version < "4.0"
waitress==2.1.2 ; python_version >= "3.8" and python_version < "4.0"
websocket-client==1.4.2 ; python_version >= "3.8" and python_version < "4.0"
werkzeug==2.2.2 ; python_version >= "3.8" and python_version < "4.0"
yarl==1.8.2 ; python_version >= "3.8" and python_version < "4.0"
zipp==3.11.0 ; python_version >= "3.8" and python_version < "3.10"

